docker/webapp/becotrack/requirements.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.11.0
mysqlclient==2.1.0
pytz
sqlparse==0.4.2
django-filter
